requirements.txt
streamlit==1.21.0
pandas==1.5.3
numpy==1.23.5
PyYAML==6.0
plotly==5.15.0
torch==2.4.0
transformers==4.45.1
bitsandbytes==0.44.1
tokenizers==0.20.0
datasets==2.12.0
peft==0.13.0
scikit-learn==1.2.2
scikit-multilearn==0.2.0
tqdm==4.64.1